% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RH.R
\name{calc_RH}
\alias{calc_RH}
\title{calc_RH}
\usage{
calc_RH(dry_bulb, wet_bulb)
}
\arguments{
\item{dry_bulb}{Dry bulb temperature in °C.}

\item{wet_bulb}{Wet bulb temperature in °C.}
}
\value{
Relative humidity as a percentage.
}
\description{
Calculate relative humidity from dry bulb temperature and wet bulb temperature.
}
\examples{
calc_RH(dry_bulb = 24, wet_bulb = 22.5)
}

