% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SVP.R
\name{calc_SVP}
\alias{calc_SVP}
\title{calc_SVP}
\usage{
calc_SVP(temp)
}
\arguments{
\item{temp}{Temperature in °C.}
}
\value{
Saturated vapour pressure in pascals.
}
\description{
Calculate saturated vapour pressure from temperature.
}
\examples{
calc_SVP(temp = 24)
}

