% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VPD.R
\name{calc_VPD}
\alias{calc_VPD}
\title{calc_VPD}
\usage{
calc_VPD(temp, rel_humidity)
}
\arguments{
\item{temp}{Temperature in °C.}

\item{rel_humidity}{Relative humidity as a percentage.}
}
\value{
Vapour pressure deficit in pascals.
}
\description{
Calculate vapour pressure deficit from temperature and relative humidity.
}
\examples{
calc_VPD(temp = 24, rel_humidity = 95)
}

