% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WD.R
\name{calc_WD}
\alias{calc_WD}
\title{calc_WD}
\usage{
calc_WD(temp)
}
\arguments{
\item{temp}{Temperature in °C.}
}
\value{
Water density in kg/m³
}
\description{
Calculate water density from temperature.
}
\examples{
calc_WD(temp = 24)
}

